(refactor) rm balances example, add value calcs to wallets example
